pyproject.toml
[project]
name = "rl-chess"
version = "0.1.0"
description = "Teaching a 7B model to beat GPT-4o at chess with GRPO"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.0.0",
    "datasets>=3.0.0",
    "openai>=1.0.0",
    "peft>=0.13.0",
    "python-chess>=1.999",
    "torch>=2.4.0",
    "transformers>=4.45.0,<5",
    "trl>=0.15.0",
    "wandb>=0.18.0",
    "vllm>=0.7.0; sys_platform == 'linux'",
    "bitsandbytes>=0.44.0; sys_platform == 'linux'",
]

[dependency-groups]
dev = [
    "mypy",
    "ruff",
]
